more BG work, starting work in C#
got the image closer to where it needs to be, still working on objects
though.

Ive decided it would be fun to also learn a new language while working
on this, so I’ve begun replicating it in C#. this means that updates
may be fairly unfrequent until I get the C# version up to speed, which
will probably take a while. I’ll still work through problems in the
java version, however.